Moving is_dask_array to utils
